ringo_kai_dev_package/backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
python-dotenv==1.0.1
supabase==2.5.0
supabase-py==2.4.4
openai==1.52.1
Pillow==11.0.0
python-multipart==0.0.12
tenacity==8.5.0
psycopg[binary]==3.1.18
